+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-## Version 1.6.3
+## Version 1.6.4
 
 ### Bugfix
-- Fixed fields SysctlParams and ACL models RecordLB in ca/lb/models and models RecordLB and ItemLBList in cb/lb/models
+- Fixed type field DiskSizeMax in Resource model in cloudapi/account/models. Rename field ItemResourceConsumption model and rename json's tags in cloudapi/account/models
+- Rename field in ListPCIDeviceRequest model in cloudapi/compute/list_pci_device
+- Delete excess field in ListDeletedRequest model in cloudapi/k8s/list_deleted
+- Add fields Page and Size in AffinityGroupsListRequest model in cloudapi/lb/affinity_groups_list
+- Fixed type field Data in ListAffinityGroups model in cloudapi/lb/models
+- Rename in ItemResourceConsumption model json's tags in cloudapi/rg/models